+++ example/sleep.ts
@@ -1,0 +1,10 @@
+const sleep = (time: number) => new Promise(resolve => setTimeout(resolve, time))
+
+Bun.serve({
+	port: 3000,
+	fetch: async () => {
+		await sleep(1000)
+
+		return new Response('Hi')
+	}
+})